Implementation of DiffCommandStep
