+++ liquibase-core/src/main/java/liquibase/command/core/DbUrlConnectionCommandStep.java
@@ -1,30 +1,19 @@
 package liquibase.command.core;
 
-import liquibase.GlobalConfiguration;
-import liquibase.Scope;
 import liquibase.command.*;
 import liquibase.configuration.ConfigurationValueObfuscator;
 import liquibase.database.Database;
-import liquibase.database.DatabaseFactory;
-import liquibase.database.core.DatabaseUtils;
 import liquibase.exception.DatabaseException;
-import liquibase.integration.commandline.LiquibaseCommandLineConfiguration;
-import liquibase.resource.ResourceAccessor;
-import liquibase.util.StringUtil;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.ResourceBundle;
-
-import static java.util.ResourceBundle.getBundle;
 
 /**
  * Internal command step to be used on CommandStep pipeline to manage the database connection.
  */
-public class DbUrlConnectionCommandStep extends AbstractCommandStep implements CleanUpCommandStep {
+public class DbUrlConnectionCommandStep extends AbstractDatabaseConnectionCommandStep implements CleanUpCommandStep {
 
     protected static final String[] COMMAND_NAME = {"dbUrlConnectionCommandStep"};
-    private static final ResourceBundle coreBundle = getBundle("liquibase/i18n/liquibase-core");
 
     public static final CommandArgumentDefinition<Database> DATABASE_ARG;
     public static final CommandArgumentDefinition<String> URL_ARG;
@@ -35,9 +24,6 @@ public class DbUrlConnectionCommandStep extends AbstractCommandStep implements C
     public static final CommandArgumentDefinition<String> DRIVER_ARG;
     public static final CommandArgumentDefinition<String> DRIVER_PROPERTIES_FILE_ARG;
 
-    public static final int ORDER = 100;
-
-    private Database database;
 
     static {
         CommandBuilder builder = new CommandBuilder(COMMAND_NAME);
@@ -87,93 +73,11 @@ public class DbUrlConnectionCommandStep extends AbstractCommandStep implements C
             String defaultCatalogName = commandScope.getArgumentValue(DEFAULT_CATALOG_NAME_ARG);
             String driver = commandScope.getArgumentValue(DRIVER_ARG);
             String driverPropertiesFile = commandScope.getArgumentValue(DRIVER_PROPERTIES_FILE_ARG);
-            this.database = createDatabaseObject(url, username, password, defaultSchemaName, defaultCatalogName, driver, driverPropertiesFile);
-            return this.database;
+            return createDatabaseObject(url, username, password, defaultSchemaName, defaultCatalogName, driver, driverPropertiesFile);
         } else {
-            return commandScope.getArgumentValue(DATABASE_ARG);
+            this.database = commandScope.getArgumentValue(DATABASE_ARG);
+            return this.database;
         }
-    }
-
-    @SuppressWarnings("java:S2095")
-    /**
-     *
-     * Method to create a Database object given these parameters
-     *
-     * @param  url                       URL to connect to
-     * @param  username                  Username credential
-     * @param  password                  Password credential
-     * @param  defaultSchemaName         Default schema for connection
-     * @param  defaultCatalogName        Default catalog for connection
-     * @param  driver                    Driver class
-     * @param  driverPropertiesFile      Additional driver properties
-     * @throws DatabaseException         Thrown when there is a connection error
-     *
-     */
-    private Database createDatabaseObject(String url,
-                                        String username,
-                                        String password,
-                                        String defaultSchemaName,
-                                        String defaultCatalogName,
-                                        String driver,
-                                        String driverPropertiesFile)
-            throws DatabaseException {
-        ResourceAccessor resourceAccessor = Scope.getCurrentScope().getResourceAccessor();
-        String databaseClassName = null;
-        Class<?> databaseClass = LiquibaseCommandLineConfiguration.DATABASE_CLASS.getCurrentValue();
-        if (databaseClass != null) {
-            databaseClassName = databaseClass.getCanonicalName();
-        }
-        String propertyProviderClass = null;
-        Class<?> clazz = LiquibaseCommandLineConfiguration.PROPERTY_PROVIDER_CLASS.getCurrentValue();
-        if (clazz != null) {
-            propertyProviderClass = clazz.getName();
-        }
-        String liquibaseCatalogName = StringUtil.trimToNull(GlobalConfiguration.LIQUIBASE_CATALOG_NAME.getCurrentValue());
-        String liquibaseSchemaName = StringUtil.trimToNull(GlobalConfiguration.LIQUIBASE_SCHEMA_NAME.getCurrentValue());
-        String databaseChangeLogTablespaceName = StringUtil.trimToNull(GlobalConfiguration.LIQUIBASE_TABLESPACE_NAME.getCurrentValue());
-        String databaseChangeLogLockTableName = StringUtil.trimToNull(GlobalConfiguration.DATABASECHANGELOGLOCK_TABLE_NAME.getCurrentValue());
-        String databaseChangeLogTableName = StringUtil.trimToNull(GlobalConfiguration.DATABASECHANGELOG_TABLE_NAME.getCurrentValue());
-
-        try {
-            defaultCatalogName = StringUtil.trimToNull(defaultCatalogName);
-            defaultSchemaName = StringUtil.trimToNull(defaultSchemaName);
-
-            database = DatabaseFactory.getInstance().openDatabase(url, username, password, driver,
-                    databaseClassName, driverPropertiesFile, propertyProviderClass, resourceAccessor);
-
-            if (!database.supportsSchemas()) {
-                if ((defaultSchemaName != null) && (defaultCatalogName == null)) {
-                    defaultCatalogName = defaultSchemaName;
-                }
-                if ((liquibaseSchemaName != null) && (liquibaseCatalogName == null)) {
-                    liquibaseCatalogName = liquibaseSchemaName;
-                }
-            }
-
-            defaultCatalogName = StringUtil.trimToNull(defaultCatalogName);
-            defaultSchemaName = StringUtil.trimToNull(defaultSchemaName);
-
-            database.setDefaultCatalogName(defaultCatalogName);
-            database.setDefaultSchemaName(defaultSchemaName);
-            database.setOutputDefaultCatalog(true);
-            database.setOutputDefaultSchema(true);
-            database.setLiquibaseCatalogName(liquibaseCatalogName);
-            database.setLiquibaseTablespaceName(databaseChangeLogTablespaceName);
-            database.setLiquibaseSchemaName(liquibaseSchemaName);
-            if (databaseChangeLogTableName != null) {
-                database.setDatabaseChangeLogTableName(databaseChangeLogTableName);
-                if (databaseChangeLogLockTableName != null) {
-                    database.setDatabaseChangeLogLockTableName(databaseChangeLogLockTableName);
-                } else {
-                    database.setDatabaseChangeLogLockTableName(databaseChangeLogTableName + "LOCK");
-                }
-            }
-            DatabaseUtils.initializeDatabase(defaultCatalogName, defaultSchemaName, database);
-        } catch (Exception e) {
-            throw new DatabaseException(e);
-        }
-
-        return database;
     }
 
     @Override
@@ -181,22 +85,4 @@ public class DbUrlConnectionCommandStep extends AbstractCommandStep implements C
         return new String[][] { COMMAND_NAME };
     }
 
-    @Override
-    public void adjustCommandDefinition(CommandDefinition commandDefinition) {
-        if (commandDefinition.getPipeline().size() == 1) {
-            commandDefinition.setInternal(true);
-        }
-    }
-
-    @Override
-    public void cleanUp(CommandResultsBuilder resultsBuilder) {
-        if (database != null) {
-            try {
-                database.close();
-                database = null;
-            } catch (Exception e) {
-                Scope.getCurrentScope().getLog(getClass()).warning(coreBundle.getString("problem.closing.connection"), e);
-            }
-        }
-    }
 }
