+++ liquibase-core/src/main/java/liquibase/command/core/ReferenceDbUrlConnectionCommandStep.java
@@ -1,0 +1,90 @@
+package liquibase.command.core;
+
+import liquibase.command.*;
+import liquibase.command.providers.ReferenceDatabase;
+import liquibase.configuration.ConfigurationValueObfuscator;
+import liquibase.database.Database;
+import liquibase.exception.DatabaseException;
+import liquibase.integration.commandline.Main;
+
+import java.util.Collections;
+import java.util.List;
+
+/**
+ * Internal command step to be used on CommandStep pipeline to manage the database connection.
+ */
+public class ReferenceDbUrlConnectionCommandStep extends AbstractDatabaseConnectionCommandStep implements CleanUpCommandStep {
+
+    protected static final String[] COMMAND_NAME = {"dbUrlConnectionCommandStep"};
+
+    public static final CommandArgumentDefinition<Database> REFERENCE_DATABASE_ARG;
+    public static final CommandArgumentDefinition<String> REFERENCE_USERNAME_ARG;
+    public static final CommandArgumentDefinition<String> REFERENCE_PASSWORD_ARG;
+    public static final CommandArgumentDefinition<String> REFERENCE_URL_ARG;
+    public static final CommandArgumentDefinition<String> REFERENCE_DEFAULT_SCHEMA_NAME_ARG;
+    public static final CommandArgumentDefinition<String> REFERENCE_DEFAULT_CATALOG_NAME_ARG;
+    public static final CommandArgumentDefinition<String> REFERENCE_DRIVER_ARG;
+    public static final CommandArgumentDefinition<String> REFERENCE_DRIVER_PROPERTIES_FILE_ARG;
+
+
+    static {
+        CommandBuilder builder = new CommandBuilder(COMMAND_NAME);
+        REFERENCE_DEFAULT_SCHEMA_NAME_ARG = builder.argument("defaultSchemaName", String.class)
+                .description("The default schema name to use for the reference database connection").build();
+        REFERENCE_DEFAULT_CATALOG_NAME_ARG = builder.argument("defaultCatalogName", String.class)
+                .description("The default catalog name to use for the reference database connection").build();
+        REFERENCE_DRIVER_ARG = builder.argument("driver", String.class)
+                .description("The JDBC driver class for the reference database").build();
+        REFERENCE_DRIVER_PROPERTIES_FILE_ARG = builder.argument("driverPropertiesFile", String.class)
+                .description("The JDBC driver properties file for the reference database").build();
+        REFERENCE_USERNAME_ARG = builder.argument(CommonArgumentNames.USERNAME, String.class)
+                .description("The reference database username").build();
+        REFERENCE_PASSWORD_ARG = builder.argument(CommonArgumentNames.PASSWORD, String.class)
+                .description("The reference database password")
+                .setValueObfuscator(ConfigurationValueObfuscator.STANDARD)
+                .build();
+        REFERENCE_DATABASE_ARG = builder.argument("database", Database.class).hidden().build();
+        REFERENCE_URL_ARG = builder.argument(CommonArgumentNames.URL, String.class).required().supersededBy(REFERENCE_DATABASE_ARG)
+                .description("The JDBC reference database connection URL").build();
+        REFERENCE_DATABASE_ARG.setSupersededBy(REFERENCE_URL_ARG);
+    }
+
+    @Override
+    public List<Class<?>> providedDependencies() {
+        return Collections.singletonList(ReferenceDatabase.class);
+    }
+
+    @Override
+    public void run(CommandResultsBuilder resultsBuilder) throws Exception {
+        CommandScope commandScope = resultsBuilder.getCommandScope();
+        commandScope.provideDependency(Database.class, this.obtainDatabase(commandScope));
+    }
+
+    /**
+     * Try to retrieve and set the database object from the command scope, otherwise creates a new one .
+     *
+     * @param commandScope current command scope
+     * @throws DatabaseException Thrown when there is a connection error
+     */
+    private Database obtainDatabase(CommandScope commandScope) throws DatabaseException {
+        if (commandScope.getArgumentValue(REFERENCE_DATABASE_ARG) == null) {
+            String url = commandScope.getArgumentValue(REFERENCE_URL_ARG);
+            String username = commandScope.getArgumentValue(REFERENCE_USERNAME_ARG);
+            String password = commandScope.getArgumentValue(REFERENCE_PASSWORD_ARG);
+            String defaultSchemaName = commandScope.getArgumentValue(REFERENCE_DEFAULT_SCHEMA_NAME_ARG);
+            String defaultCatalogName = commandScope.getArgumentValue(REFERENCE_DEFAULT_CATALOG_NAME_ARG);
+            String driver = commandScope.getArgumentValue(REFERENCE_DRIVER_ARG);
+            String driverPropertiesFile = commandScope.getArgumentValue(REFERENCE_DRIVER_PROPERTIES_FILE_ARG);
+            return createDatabaseObject(url, username, password, defaultSchemaName, defaultCatalogName, driver, driverPropertiesFile);
+        } else {
+            this.database = commandScope.getArgumentValue(REFERENCE_DATABASE_ARG);
+            return this.database;
+        }
+    }
+
+    @Override
+    public String[][] defineCommandNames() {
+        return new String[][] { COMMAND_NAME };
+    }
+
+}
