+++ liquibase-core/src/main/java/liquibase/command/providers/ReferenceDatabase.java
@@ -1,0 +1,7 @@
+package liquibase.command.providers;
+
+/**
+ * Marker interface for Reference Database providers used by CommandStep Framework.
+ */
+public interface ReferenceDatabase {
+}
